backend/requirements.txt
fastapi
uvicorn
tensorflow==1.14.0
keras==2.2.4
scikit-learn==0.20.2
docopt
numpy
PyJWT
passlib[bcrypt]
# Pin bcrypt to a version compatible with passlib 1.7.x to avoid __about__ warning
bcrypt==4.0.1
